functions/process_upload/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.17.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
